update: コマンドライン引数周りを実装

+++ main.go
@@ -1,21 +1,47 @@
 package main
 
 import (
+	"bufio"
 	"flag"
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 )
 
 func main() {
+	var (
+		force bool
+		src   string
+		dest  string
+	)
+
+	wd, err := os.Getwd()
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "error while parsing args: %v\n", err)
+		return
+	}
+
+	flag.StringVar(&src, "i", "", "a input yaml file (required)")
+	flag.StringVar(&dest, "o", filepath.Join(wd, "surfaces.txt"), "an output file")
+	flag.BoolVar(&force, "f", false, "skip overwriting confirmation")
 	flag.Parse()
 
-	if flag.NArg() == 0 {
+	info, err := os.Stat(dest)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "error while parsing args: %v\n", err)
+		return
+	} else if info.IsDir() {
+		fmt.Fprintln(os.Stderr, "error while parsing args:", dest, "is directory")
+		return
+	}
+
+	if src == "" {
 		flag.Usage()
 		return
 	}
 
-	data, err := loadYaml(flag.Arg(0))
+	data, err := loadYaml(src)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error while loading yaml: %v\n", err)
 		return
@@ -27,12 +53,18 @@ func main() {
 
 	result := formatSurfaces(data, surfaces, surfaceList)
 
-	execPath, err := os.Executable()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "error while writing surfaces.txt: %v\n", err)
+	if !force && isFileExists(dest) {
+		fmt.Println("overwrite", dest+"? (y/n)")
+		scanner := bufio.NewScanner(os.Stdin)
+		for scanner.Scan() {
+			if strings.Contains(scanner.Text(), "n") {
+				return
+			} else if strings.Contains(scanner.Text(), "y") {
+				break
+			}
+		}
 	}
 
-	dest := filepath.Join(filepath.Dir(execPath), "surfaces.txt")
 	if err := os.WriteFile(dest, []byte(result), 0755); err != nil {
 		fmt.Fprintf(os.Stderr, "error while writing surfaces.txt: %v\n", err)
 		return
